Revert addition of new public APIs and creation of named configurations
* As requested in https://github.com/gradle/gradle/pull/28198
* Unfixes: https://github.com/gradle/gradle/issues/27309
* Unfixes: https://github.com/gradle/gradle/issues/18435
