+++ platforms/jvm/scala/src/main/java/org/gradle/api/tasks/scala/ScalaDoc.java
@@ -34,7 +34,7 @@ import org.gradle.api.tasks.Optional;
 import org.gradle.api.tasks.OutputDirectory;
 import org.gradle.api.tasks.PathSensitive;
 import org.gradle.api.tasks.PathSensitivity;
-import org.gradle.api.tasks.ScalaJar;
+import org.gradle.api.tasks.scala.internal.ScalaJar;
 import org.gradle.api.tasks.SourceTask;
 import org.gradle.api.tasks.TaskAction;
 import org.gradle.api.tasks.scala.internal.GenerateScaladoc;
